Apply namespaces to attributes immediately.
Namespace declarations take effect immediately in XML affecting the
current element and attributes. The current implementation did not
appear to apply namespaces to the attributes belonging to the same
element in which the namespace was declared.

Upon inspecting the code, the loop that is supposed to perform this
assignment does not actually loop through the attribute array. It
increments a index, but it does update the local variable used to
reference the attribute in the attribute list at the index.

With this commit, the local variable is updated. A test has been added
to the DOM attributes test suite to ensure that namespaces are applied
to attributes immediately.