Fix insert of empty DocumentFragment.
When inserting an empty `DocumentFragment` a null refererence exception
is thrown. I've edited the internal `insertBefore` implementation to
detect an empty `DocumentFragment` and return early.

To test this insertion, I created a new test script for
`DocumentFragment` tests. If there are future tests of
`DocumentFragment` specific logic, then can go there.

Updated the `bigeasy` contributor entry to the full name of said
contributor.